move files to subdirectory and update imports
