+++ parker_dycops2022/parker_dycops2022/common/serialize/interpolate.py
@@ -2,7 +2,7 @@ import itertools
 from scipy import interpolate
 import numpy as np
 
-from common.serialize.data_from_model import (
+from parker_dycops2022.common.serialize.data_from_model import (
         apply_function_elementwise,
         _get_structured_variable_data_from_dict,
         )
