+++ parker_dycops2022/parker_dycops2022/plot_discretization_errors.py
