+++ parker_dycops2022/parker_dycops2022/scaling_factor.json
